# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SOC Analyzer",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/dashboard": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Get Dashboard",
        "operationId": "get_dashboard_api_dashboard_get",
        "parameters": [
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dashboard"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard/review": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Get Review",
        "description": "The latest company-wide review (null if none was written yet).",
        "operationId": "get_review_api_dashboard_review_get",
        "parameters": [
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ReviewOut"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Review Api Dashboard Review Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "dashboard"
        ],
        "summary": "Create Review",
        "description": "Write a review of the company-wide picture now (Claude if configured, else the template).\nPlain `def`: the API call (~20 s) runs in the thread pool, not on the event loop.",
        "operationId": "create_review_api_dashboard_review_post",
        "parameters": [
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CategoryCount": {
        "properties": {
          "category": {
            "type": "string",
            "title": "Category"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "incidents": {
            "type": "integer",
            "title": "Incidents"
          }
        },
        "type": "object",
        "required": [
          "category",
          "label",
          "incidents"
        ],
        "title": "CategoryCount"
      },
...
      "Dashboard": {
        "properties": {
          "kpis": {
            "$ref": "#/components/schemas/Kpis"
          },
          "datasets": {
            "items": {
              "$ref": "#/components/schemas/DatasetCount"
            },
            "type": "array",
            "title": "Datasets"
          },
          "users_at_risk": {
            "items": {
              "$ref": "#/components/schemas/RiskyUser"
            },
            "type": "array",
            "title": "Users At Risk"
          },
          "severity": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Severity"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/CategoryCount"
            },
            "type": "array",
            "title": "Categories"
          },
          "top_incidents": {
            "items": {
              "$ref": "#/components/schemas/DashboardIncident"
            },
            "type": "array",
            "title": "Top Incidents"
          },
          "top_entities": {
            "items": {
              "$ref": "#/components/schemas/Entity"
            },
            "type": "array",
            "title": "Top Entities"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/DayCount"
            },
            "type": "array",
            "title": "Timeline"
          }
        },
        "type": "object",
        "required": [
          "kpis",
          "datasets",
          "users_at_risk",
          "severity",
          "categories",
          "top_incidents",
          "top_entities",
          "timeline"
        ],
        "title": "Dashboard"
      },
...
      "DatasetCount": {
        "properties": {
          "upload_id": {
            "type": "integer",
            "title": "Upload Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "first_event": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Event"
          },
          "last_event": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Event"
          },
          "events": {
            "type": "integer",
            "title": "Events"
          },
          "flagged": {
            "type": "integer",
            "title": "Flagged"
          },
          "blocked": {
            "type": "integer",
            "title": "Blocked"
          }
        },
        "type": "object",
        "required": [
          "upload_id",
          "filename",
          "first_event",
          "last_event",
          "events",
          "flagged",
          "blocked"
        ],
        "title": "DatasetCount"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Kpis": {
        "properties": {
          "incidents": {
            "type": "integer",
            "title": "Incidents"
          },
          "critical": {
            "type": "integer",
            "title": "Critical"
          },
          "affected_users": {
            "type": "integer",
            "title": "Affected Users"
          },
          "flagged_events": {
            "type": "integer",
            "title": "Flagged Events"
          },
          "blocked_events": {
            "type": "integer",
            "title": "Blocked Events"
          },
          "events": {
            "type": "integer",
            "title": "Events"
          },
          "datasets": {
            "type": "integer",
            "title": "Datasets"
          },
          "uploads": {
            "type": "integer",
            "title": "Uploads"
          },
          "findings": {
            "type": "integer",
            "title": "Findings"
          },
          "findings_high": {
            "type": "integer",
            "title": "Findings High"
          },
          "open_cases": {
            "type": "integer",
            "title": "Open Cases"
          },
          "unassigned_urgent": {
            "type": "integer",
            "title": "Unassigned Urgent"
          },
          "high_risk_users": {
            "type": "integer",
            "title": "High Risk Users"
          },
          "high_risk_ips": {
            "type": "integer",
            "title": "High Risk Ips"
          },
          "events_last_period": {
            "type": "integer",
            "title": "Events Last Period"
          },
          "events_previous_period": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Events Previous Period"
          },
          "period_days": {
            "type": "integer",
            "title": "Period Days"
          }
        },
        "type": "object",
        "required": [
          "incidents",
          "critical",
          "affected_users",
          "flagged_events",
          "blocked_events",
          "events",
          "datasets",
          "uploads",
          "findings",
          "findings_high",
          "open_cases",
          "unassigned_urgent",
          "high_risk_users",
          "high_risk_ips",
          "events_last_period",
          "events_previous_period",
          "period_days"
        ],
        "title": "Kpis"
      },
...
      "ReviewOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "source": {
            "type": "string",
            "enum": [
              "ai",
              "template"
            ],
            "title": "Source"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "headline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headline"
          },
          "key_findings": {
            "items": {
              "$ref": "#/components/schemas/KeyFinding"
            },
            "type": "array",
            "title": "Key Findings"
          },
          "actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Actions"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "incident_count": {
            "type": "integer",
            "title": "Incident Count"
          },
          "stale": {
            "type": "boolean",
            "title": "Stale"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source",
          "model",
          "headline",
          "key_findings",
          "actions",
          "summary",
          "created_at",
          "created_by",
          "incident_count",
          "stale"
        ],
        "title": "ReviewOut"
      },
...
      "RiskyUser": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "priority": {
            "type": "string",
            "enum": [
              "critical",
              "high",
              "medium",
              "low"
            ],
            "title": "Priority"
          },
          "priority_score": {
            "type": "number",
            "title": "Priority Score"
          },
          "incidents": {
            "type": "integer",
            "title": "Incidents"
          },
          "findings": {
            "type": "integer",
            "title": "Findings"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "upload_id": {
            "type": "integer",
            "title": "Upload Id"
          }
        },
        "type": "object",
        "required": [
          "username",
          "priority",
          "priority_score",
          "incidents",
          "findings",
          "reason",
          "upload_id"
        ],
        "title": "RiskyUser"
      },
...
    }
  }
}